backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
bcrypt
pydantic
python-multipart
